backened/requirements.txt
flask
flask-cors
tensorflow-cpu
numpy
pandas
requests
scikit-learn
gunicorn
